# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    GenericError:
        description: GenericError GenericError is a generic error message when the product path is not valid
        properties:
            message:
                description: message
                type: string
                x-go-name: Message
        type: object
        x-go-package: micro-services/client/models
    Product:
        description: Product Product define the structure for an API product
        properties:
            description:
                description: the dedcription for this product
                maxLength: 1000
                type: string
                x-go-name: Description
            id:
                description: the id for this user
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                description: the name for this user
                minLength: 3
                type: string
                x-go-name: Name
            price:
                description: the name for this user
                format: double
                minimum: 0.1
                type: number
                x-go-name: Price
            sku:
                description: the name for this user
                pattern: '[a-z]+-[a-z]+-[a-z]+'
                type: string
                x-go-name: SKU
        required:
            - id
            - name
            - description
            - price
            - sku
        type: object
        x-go-package: micro-services/client/models
    ValidationError:
        description: ValidationError ValidationError is a collection of validation error messages
        properties:
            Messages:
                description: messages
                items:
                    type: string
                type: array
        type: object
        x-go-package: micro-services/client/models
info:
    description: Documentation for Product API222
    title: of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            parameters:
                - description: |-
                    Currency used when returning the price of the product,
                    when not specified currency is returned in GBP.
                  in: query
                  name: Currency
                  type: string
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
        post:
            description: Create a new product
            operationId: createProdcut
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
                "422":
                    $ref: '#/responses/errorValidation'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
        put:
            description: Create a new product
            operationId: updateProduct
            parameters:
                - description: |-
                    Product data structure to Update or Create.
                    Note: the id field is ignored by update and create operations
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/Product'
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "433":
                    $ref: '#/responses/errorValidation'
            tags:
                - products
    /products/{id}:
        delete:
            description: Update a products details
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from the database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
                "404":
                    $ref: '#/responses/errorResponse'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
        get:
            description: Returns a single product
            operationId: listSingleProduct
            parameters:
                - description: |-
                    Currency used when returning the price of the product,
                    when not specified currency is returned in GBP.
                  in: query
                  name: Currency
                  type: string
                - description: The id of the product for which the operation relates
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
                "404":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
produces:
    - application/json
responses:
    errorResponse:
        description: Generic error message returned as a string
        schema:
            $ref: '#/definitions/GenericError'
    errorValidation:
        description: Validation errors defined as an array of strings
        schema:
            $ref: '#/definitions/ValidationError'
    noContent:
        description: ""
    noContentResponse:
        description: No content is returned by this API endpoint
    productResponse:
        description: Data structure representing a single product
        schema:
            $ref: '#/definitions/Product'
    productsResponse:
        description: A list of products
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
schemes:
    - http
swagger: "2.0"